Accept all values for tvshow status.
Previously this addon dropped all but "Ended". This removes that
restriction. When paired with a recent build of Sonarr (v4), the user
is supplied additional information such as if a show is "Continuing".

(Using the default Kodi skin, this information shows up on the "info"
screen for a TV show.)